move swim tests to swim package
