+++ keeper/src/test/scala/zio/keeper/membership/swim/NodesSpec.scala
@@ -1,16 +1,16 @@
-package zio.keeper.membership
+package zio.keeper.membership.swim
 
 import zio.ZLayer
 import zio.clock.Clock
-import zio.keeper.NodeAddress
-import zio.keeper.membership.swim.Nodes
+import zio.keeper.membership.MembershipEvent
+import zio.keeper.membership.swim.Nodes._
+import zio.keeper.{ KeeperSpec, NodeAddress }
 import zio.logging.Logging
 import zio.stream.Sink
 import zio.test.Assertion._
-import zio.test.{ DefaultRunnableSpec, _ }
-import zio.keeper.membership.swim.Nodes._
+import zio.test._
 
-object NodesSpec extends DefaultRunnableSpec {
+object NodesSpec extends KeeperSpec {
 
   val logger = Logging.console((_, line) => line)
