+++ keeper/src/test/scala/zio/keeper/membership/swim/ProtocolSpec.scala
@@ -1,14 +1,13 @@
-package zio.keeper.membership
+package zio.keeper.membership.swim
 
 import zio.ZIO
-import zio.keeper.{ ByteCodec, NodeAddress }
-import zio.keeper.membership.PingPong.{ Ping, Pong }
-import zio.keeper.membership.swim.{ ConversationId, Message, Protocol }
+import zio.keeper.membership.swim.PingPong.{ Ping, Pong }
+import zio.keeper.{ ByteCodec, KeeperSpec, NodeAddress }
 import zio.stream.ZStream
 import zio.test.Assertion._
 import zio.test._
 
-object ProtocolSpec extends DefaultRunnableSpec {
+object ProtocolSpec extends KeeperSpec {
 
   val protocolDefinition = Protocol[PingPong].make(
     {
